backend/requirements.txt
cloudinary==1.44.1
Django==5.2.9
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pillow==12.0.0
python-dotenv==1.2.1
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
psycopg2-binary==2.9.11
